Split PasswordResetPage into subcomponents
